requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
beautifulsoup4==4.13.4
boto3==1.38.32
bs4==0.0.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
extra-streamlit-components==0.1.80
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-openai==0.3.17
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langchain-together==0.3.0
markdown-it-py==3.0.0
numpy==2.2.6
openai==1.79.0
pandas==2.2.3
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
PyAutoGUI==0.9.54
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pygame==2.6.1
PyMuPDF==1.26.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.32.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
streamlit==1.45.1
streamlit-chat==0.1.1
streamlit-cookies-manager==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
